\name{bitrugs-internal}
\alias{network_mcmc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal functions
}
\description{
Used internally only
}
\usage{
network_mcmc(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{
Arguments taken by network_mcmc
}
}
\keyword{internal}
\details{
C function driving the MCMC algorithm.
}

