\name{bitrugs}
\alias{bitrugs}
\docType{package}
\title{
Bayesian Inference of Transmission Routes Using Genome Sequences
}
\description{
MCMC methods to estimate transmission dynamics and infection routes in hospitals using genomic sampling data.
}
\details{
\tabular{ll}{
Package: \tab bitrugs\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2016-05-29\cr
License: \tab GPL-3\cr
}
}
\author{
Colin Worby <cworby@hsph.harvard.edu>
}
\references{
Worby CJ et al. (2016) Reconstructing transmission networks for communicable diseases using densely sampled genomic data: a generalized approach. Annals of Applied Statistics.
}

\keyword{Nosocomial infection, epidemiology, MCMC, genome sequence, transmission tree}
